+++ src/pages/dashboard/index.tsx
@@ -1,26 +1,24 @@
 import AppLayout from '@shared/layouts/app-layout';
 import type { NextPage } from 'next';
 import { MOCK_CARDS } from '../../mock-data';
-import ProjectGroup, { ProjectCategory } from './project-group';
-import { useMemo } from 'react';
+import ProjectGroup from '../../feature/dashboard/project-group';
+import { ProjectCategory } from 'feature/dashboard/interfaces';
 
 const Dashboard: NextPage = () => {
   const projects = MOCK_CARDS;
 
-  const projectsSortedByCategory = useMemo(() => {
-    return projects.reduce((acc, project) => {
-      const category = project.category;
-      if (!acc[category]) {
-        acc[category] = [];
-      }
-      acc[category].push(project);
-      return acc;
-    }, {} as ProjectCategory);
-  }, [projects]);
+  const projectsGroupedByCategory = projects.reduce((acc, project) => {
+    const category = project.category;
+    if (!acc[category]) {
+      acc[category] = [];
+    }
+    acc[category].push(project);
+    return acc;
+  }, {} as ProjectCategory);
 
   return (
     <AppLayout>
-      <ProjectGroup projects={projectsSortedByCategory} />
+      <ProjectGroup projects={projectsGroupedByCategory} />
     </AppLayout>
   );
 };
